+++ cloud/oauth-server/src/main/kotlin/pers/acp/admin/oauth/controller/api/RoleController.kt
@@ -27,7 +27,6 @@ import pers.acp.spring.cloud.annotation.AcpCloudDuplicateSubmission
 
 import javax.annotation.PostConstruct
 import javax.validation.Valid
-import javax.validation.constraints.NotBlank
 import javax.validation.constraints.NotEmpty
 import javax.validation.constraints.NotNull
 
@@ -79,7 +78,7 @@ constructor(private val logAdapter: LogAdapter,
     @PreAuthorize(RoleConfigExpression.roleQuery)
     @GetMapping(value = [OauthApi.roleConfig], produces = [MediaType.APPLICATION_JSON_VALUE])
     @Throws(ServerException::class)
-    fun roleList(user: OAuth2Authentication): ResponseEntity<List<Role>> = ResponseEntity.ok(roleDomain.getRoleList(user))
+    fun roleList(): ResponseEntity<List<Role>> = ResponseEntity.ok(roleDomain.getRoleList())
 
     @ApiOperation(value = "新建角色信息", notes = "名称、编码、应用ID、级别、序号、关联用户、关联菜单、关联模块功能")
     @ApiResponses(ApiResponse(code = 201, message = "创建成功", response = Role::class), ApiResponse(code = 400, message = "参数校验不通过；角色编码非法，请重新输入；", response = ErrorVo::class))